feat: add modal for new transaction
Implements a modal to add a new transaction
