+++ src/pages/Home/index.jsx
@@ -1,18 +1,9 @@
 import { useEffect } from "react";
 import { useState } from "react";
-import { v4 as uuidv4 } from "uuid";
+import ModalAddTransaction from "../../components/ModalAddTransaction";
 
 function Home() {
   const [transactions, setTransactions] = useState([]);
-
-  const [transaction, setTransaction] = useState({
-    nome: "",
-    categoria: "",
-    data: "",
-    valor: "",
-  });
-
-  console.log(transactions);
 
   useEffect(() => {
     getTransactions();
@@ -36,32 +27,6 @@ function Home() {
       .then((response) => response.json())
       .then((data) => {
         setTransactions(data);
-      })
-      .catch((error) => {
-        console.error("Error:", error);
-      });
-  };
-
-  const submitTransaction = (event) => {
-    event.preventDefault();
-
-    const url =
-      "https://sheet2api.com/v1/rtjzbZKQ2CY1/budget-management/P%C3%A1gina1";
-
-    fetch(url, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        ...transaction,
-        id: uuidv4(),
-      }),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        setTransactions([...transactions, data]);
-        console.log("Success:", data);
       })
       .catch((error) => {
         console.error("Error:", error);
@@ -119,9 +84,11 @@ function Home() {
               <h2 className="text-lg font-semibold">Transactions</h2>
             </div>
             <div className="text-end">
-              <button className="text-start p-3 bg-blue-400 text-white text-lg font-semibold cursor-pointer">
-                Add Transaction
-              </button>
+              <ModalAddTransaction
+                handleNewTransaction={(newTransaction) =>
+                  setTransactions([...transactions, newTransaction])
+                }
+              />
             </div>
           </div>
 
@@ -166,58 +133,6 @@ function Home() {
             </ul>
           </div>
         </section>
-
-        <form>
-          <label htmlFor="nome">Nome:</label>
-          <input
-            value={transaction.nome}
-            type="text"
-            id="nome"
-            name="nome"
-            onChange={(e) =>
-              setTransaction({ ...transaction, nome: e.target.value })
-            }
-          />
-          <br />
-          <br />
-          <label htmlFor="categoria">Categoria:</label>
-          <input
-            value={transaction.categoria}
-            type="text"
-            id="categoria"
-            name="categoria"
-            onChange={(e) =>
-              setTransaction({ ...transaction, categoria: e.target.value })
-            }
-          />
-          <br />
-          <br />
-          <label htmlFor="valor">Valor:</label>
-          <input
-            value={transaction.valor}
-            type="number"
-            id="valor"
-            name="valor"
-            onChange={(e) =>
-              setTransaction({ ...transaction, valor: e.target.value })
-            }
-          />
-          <br />
-          <br />
-          <label htmlFor="data">Data:</label>
-          <input
-            value={transaction.data}
-            type="date"
-            id="data"
-            name="data"
-            onChange={(e) =>
-              setTransaction({ ...transaction, data: e.target.value })
-            }
-          />
-          <br />
-          <br />
-          <button onClick={submitTransaction}>Submit</button>
-        </form>
       </section>
     </>
   );
